Put item collection into seperate users

+++ lib/repositories/items_repository.dart
@@ -1,10 +1,17 @@
 import 'package:cantwait28/models/item_model.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 class ItemsRepository {
 
   Stream<List<ItemModel>> getItemStream() {
+  final userID = FirebaseAuth.instance.currentUser?.uid;
+  if (userID == null) {
+    throw Exception('User is not logged in');
+  }
    return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .orderBy('release_date')
         .snapshots()
@@ -20,14 +27,26 @@ class ItemsRepository {
         }
 
   Future<void> delete ({required String id}) {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+  if (userID == null) {
+    throw Exception('User is not logged in');
+  }
     return FirebaseFirestore.instance
+          .collection('users')
+          .doc(userID)
           .collection('items')
           .doc(id)
           .delete();
   }
 
   Future<ItemModel> get ({required String id}) async{
+  final userID = FirebaseAuth.instance.currentUser?.uid;
+  if (userID == null) {
+    throw Exception('User is not logged in');
+  }
    final doc = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(userID)
           .collection('items')
           .doc(id)
           .get();
@@ -44,7 +63,12 @@ class ItemsRepository {
     String imageURL, 
     DateTime releaseDate,) 
     async {
-    await FirebaseFirestore.instance.collection('items').add(
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+    throw Exception('User is not logged in');
+  }
+    await FirebaseFirestore.instance.collection('users')
+        .doc(userID).collection('items').add(
         {
           'title': title,
           'image_url': imageURL,
